cats: avoid cyclic dependency via bareoscats
bareoscats was removed, because only bareoscats-postgresql was left and
the need for the stub backend is gone.
However, bareoscats-postgresql depends on bareossql and vice versa, so
that these modules would have to be merged into one. This patch
re-introduces bareoscats so we can avoid the dependency cycle and
postpone the merge of the two modules.